@@ -11,13 +11,13 @@
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
-    "@typechain/ethers-v5": "^0.0.1",
+    "@typechain/ethers-v5": "^2.0.0",
     "@types/bn.js": "^4.11.6",
-    "ts-node": "^8.8.2",
-    "typechain": "^2.0.0",
-    "typescript": "^3.8.3"
+    "ts-node": "^9.0.0",
+    "typechain": "^3.0.0",
+    "typescript": "^4.0.5"
   },
   "dependencies": {
-    "ethers": "^5.0.3"
+    "ethers": "^5.0.19"
   }
 }
